Reworked folder structure for Home page + Reworked skeleton to match task's description model.
- Moved home from `./src/containers` to `./src/containers/home` structure. Reasoning is for `./src/container/home/images` to store relevant images belonging to home.
- Added `I am...` h1 text.
- Ditched  Floating Action Buttons (FAB) in favor of image wrapped in a button wrapped in a link to serve as Image Button Links to match model given for task
- Added image files for use in buttons.
- Added drop shadow to new buttons.
- Fiddled with flexbox layout options.
- Added Route to `/studentlogin` which got missed in the last commit...
